Fix Dithering setting saving (at last)
This issue since the very beginning has always plagued the WiiSX family (WiiSX/-R/-RX/WiiStation), so it's time to finally get rid of this problem and finally fix it.

Added a setting in settingsRX.cfg file so Dithering can be modified manually, and will be saved and loaded correctly.

Valid values for Dithering setting (Dithering =):
0 = no dithering (None)
1 = default (by default is set to No dithering in PEOPS GPU)
2 = always use dithering (Always)